// libs/platform/list/access-data/src/lib/enums/platform-list/affiliate-visibility.enum.ts
export enum AffiliateVisibilityIdEnum {
    Public = 1,
    RequireApproval = 2,
    Private = 3
}

export enum AffiliateVisibilityNameEnum {
    Public = 'public',
    RequireApproval = 'require_approval',
    Private = 'private'
}

export enum AffiliateVisibilityTranslateEnum {
    Public = 'visibility_types.public',
    RequireApproval = 'visibility_types.require',
    Private = 'visibility_types.private'
}